requirements.txt
<<<<<<< HEAD
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.5
django-allauth==0.57.0
django-cors-headers==4.2.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-yasg==1.21.7
future==0.18.3
idna==3.4
inflection==0.5.1
kombu==5.3.2
mailjet-rest==1.3.4
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.1
prompt-toolkit==3.0.39
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
upstox==2.0.3
uritemplate==4.1.1
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.9
websocket-client==1.6.4
=======
python==3.8.0
django==3.0
drf_yasg
django-rest-auth 
django-allauth 
djangorestframework-jwt
django-cors-headers
mysqlclient
djangorestframework-simplejwt
phonenumbersmailjet_rest
>>>>>>> 9760532483ae95be0ec254447b44854be07497bf
